Extract fill_tile_zeros to dataflow utils
